Makes RandomManager a proper singleton.
