+++ include/flamegpu/runtime/utility/RandomManager.cuh
@@ -4,35 +4,48 @@
 #include <curand_kernel.h>
 #include <cstdint>
 #include <random>
+#include <string>
 
 #include "flamegpu/runtime/utility/AgentRandom.cuh"
 
+#include "flamegpu/sim/Simulation.h"
+
 /**
- * Static manager for initialising simulation wide random with a common seed
+ * Singleton manager for initialising simulation wide random with a common seed
+ * This is an internal class, that should not be accessed directly by modellers
  * Manages the shared array of curand state use by agent functions
  * Manages the random engine/s used by host functions
- * Pairs with HostRandom and device side AgentRandom
+ * @see AgentRandom For random number generation during agent functions on the device
+ * @see HostRandom For random number generation during host functions
  */
 class RandomManager {
+    /**
+     * Requires instance for host function random generation
+     */
+    friend class HostRandom;
+    /**
+     * Handles seeding of random generation
+     */
+    // friend class Simulation;  // bool CUDAAgentModel::step(const Simulation&)
+    friend int Simulation::checkArgs(int, const char**, std::string &);
+    /**
+     * Calls resize() during simulation execution to resize device random array
+     */
+    friend class CUDAAgentModel;  // bool CUDAAgentModel::step(const Simulation&)
  public:
     /**
      * Inherit size_type from include-public partner class
      */
     typedef AgentRandom::size_type size_type;
     /**
-     * Utility for passing to init()
+     * Utility for generating a psuesdo-random seed to pass to init
      */
-    static uint64_t seedFromTime();
+    uint64_t seedFromTime();
     /**
-     * Acts as constructor
+     * Reseeds all owned random generators
      * @note Can be called multiple times to reseed, doing so releases existing memory allocations
      */
-    static void init(const unsigned int &seed);
-    /**
-     * Acts as destructor
-     * @note Safe to call multiple times
-     */
-    static void free();
+    void reseed(const unsigned int &seed);
     /**
      * Resizes random array according to the rules:
      *   while(length<_length)
@@ -41,81 +54,103 @@ class RandomManager {
      *     while(length*shrinkModifier>_length)
      *       length*=shrinkModifier
      */
-    static bool resize(const size_type &_length);
+    bool resize(const size_type &_length);
     /**
      * Accessors
      */
-    static void setGrowthModifier(float);
-    static float getGrowthModifier();
-    static void setShrinkModifier(float);
-    static float getShrinkModifier();
-    template<typename T, typename dist>
+    void setGrowthModifier(float);
+    float getGrowthModifier();
+    void setShrinkModifier(float);
+    float getShrinkModifier();
     /**
      * Generates a random number with the provided distribution
-     * Note: Not believed to be thread-safe!
+     * @param distribution A distribution object defined by <random>
+     * @tparam dist random distribution type to be used for generation (this should be implicitly detected)
+     * @tparam T return type
+     * @note Not believed to be thread-safe!
      */
-    static T getDistribution(dist &distribution);
+    template<typename T, typename dist>
+    T getDistribution(dist &distribution);
     /**
      * Returns length of curand state array currently allocated
      */
-    static size_type size();
-    static uint64_t seed();
-
- protected:
-    /**
-     * Protected constructor prevents the class being instantiated
-     * Device array of currand must be declared as __device__
-     * Therefore it cannot be instanced, so neither should it's manager.
-     */
-    RandomManager() {}
+    size_type size();
+    uint64_t seed();
 
  private:
     /**
      * Random seed used to initialise all currently allocated curand states
      */
-    static unsigned int mSeed;
+    unsigned int mSeed = 0;
     /**
      * Local copy of the length of d_random_state
      */
-    static size_type length;
+    size_type length = 0;
     /**
      * Minimum allocation length of d_random_state
      */
-    static const size_type min_length;
+    const size_type min_length = 1024;
     /**
      * Rate at which d_random_state's length shrinks
      * Must be in the range 0 < x <= 1.0
      * A value of 1.0 means the curand array will never decrease in size 
      */
-    static float shrinkModifier;
+    float shrinkModifier = 1.0f;
     /**
      * Rate at which d_random_state's length grows
      * Must be in greater than 1.0
      */
-    static float growthModifier;
+    float growthModifier = 1.5f;
     /**
      * Actually performs the resizing of the curand state array
      * If shrinking, 'deallocated' curand states are backed up to host until next required,
      *  this prevents them being reinitialised with the same seed.
      */
-    static void resizeDeviceArray(const size_type &_length);
+    void resizeDeviceArray(const size_type &_length);
     /**
      * Host copy of 'deallocated' curand states
      * When the device array shrinks in size, shrunk away curand states are stored here
      * @note h_max_random_state will be allocated to length h_max_random_size
      * However, it will only be initialised from hd_random_size(aka length) onwards
      */
-    static curandState *h_max_random_state;
+    curandState *h_max_random_state = nullptr;
     /**
      * Allocated length of h_max_random_state
      */
-    static RandomManager::size_type h_max_random_size;
+    size_type h_max_random_size = 0;
     /**
      * Seeded host random generator
      * Don't believe this to be thread-safe!
      * @note - std::default_random_engine is platform (compiler) specific. GCC (7.4) defaults to a linear_congruential_engine, which returns the same sequence for seeds 0 and 1. mt19937 is the default in MSVC and generally seems more sane.
      */
-    static std::mt19937 host_rng;
+    std::mt19937 host_rng;
+    /**
+     * Acts as destructor
+     * @note Safe to call multiple times
+     */
+    void free();
+    /**
+     * Remainder of class is singleton pattern
+     */
+    /**
+     * Creates the singleton and calls reseed() with the return value from seedFromTime()
+     */
+    RandomManager();
+
+ protected:
+     /**
+      * Returns the RandomManager singleton instance
+      */
+     static RandomManager& getInstance() {
+         static RandomManager instance;  // Guaranteed to be destroyed.
+         return instance;                // Instantiated on first use.
+     }
+     ~RandomManager();
+
+ public:
+    // Public deleted creates better compiler errors
+    RandomManager(RandomManager const&) = delete;
+    void operator=(RandomManager const&) = delete;
 };
